Split settingsMenu
